+++ lib/src/config_provider/config_types.dart
@@ -6,6 +6,7 @@
 import 'dart:io';
 
 import 'package:quiver/pattern.dart' as quiver;
+
 import 'path_finder.dart';
 
 class CommentType {